feedback: move helper methods to utils.js
