+++ packages/react-instantsearch-core/src/core/indexUtils.js
@@ -1,24 +1,9 @@
-import { has, omit, get } from 'lodash';
+import { get, has, omit } from 'lodash';
 
 export function getIndexId(context) {
   return context && context.multiIndexContext
     ? context.multiIndexContext.targetedIndex
     : context.ais.mainTargetedIndex;
-}
-
-export function addAbsolutePositions(hits, hitsPerPage, page) {
-  return hits.map((hit, index) => ({
-    ...hit,
-    __position: hitsPerPage * page + index + 1,
-  }));
-}
-
-export function addQueryID(hits, queryID) {
-  if (!queryID) return hits;
-  return hits.map(hit => ({
-    ...hit,
-    __queryID: queryID,
-  }));
 }
 
 export function getResults(searchResults, context) {
