+++ packages/react-instantsearch-core/src/core/utils.js
@@ -44,3 +44,18 @@ export const removeEmptyKey = obj => {
 
   return obj;
 };
+
+export function addAbsolutePositions(hits, hitsPerPage, page) {
+  return hits.map((hit, index) => ({
+    ...hit,
+    __position: hitsPerPage * page + index + 1,
+  }));
+}
+
+export function addQueryID(hits, queryID) {
+  if (!queryID) return hits;
+  return hits.map(hit => ({
+    ...hit,
+    __queryID: queryID,
+  }));
+}
